backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
factory-boy==3.2.0
Faker==8.10.1
flake8==3.9.2
inflection==0.5.1
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
